refactor(decision): tighten the results-map pin query and its filter plumbing
Narrows the shared pin read to the columns the map schema actually ships
(it previously selected both joined profiles whole, including the
generated search tsvector, over an unbounded row set), derives the
scope-input type from the wire schema instead of restating the omit
list, drops the unused `isAdmin` from the scope contract, and builds the
results tab's list filter and pin filter from one helper so a new filter
can't reach one without the other.
